Extract EditScreen composables out of CreationScreen
